+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## 1.0.0
+## 0.5.0
 
 ### Added
 
-* Submission : a password can be set to extract submitted archive
+* Status: a new command to get Detect profile status
+
+## 0.4.0
+
+### Added
+
+* Submission: a password can be set to extract submitted archive